pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fbm-auth"
version = "0.1.0"
description = "Shared authentication package for FBM Copilot Suite"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",
]

[tool.setuptools.packages.find]
include = ["fbm_auth*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
